huffman and import fixed
